This machine learning code was created in 1 shot prompt to the new o1-preview for OPENAI.
'Create a machine learning model using supervised learning for predictions and unsupervised for relations and clusters. For the data you can create it. The goal is that the model will predict the percentage of a crime to happen in Orlando. So do the EDA, modeling, testing and finally a great UI for predictions with human input. This to be run in a Jupyter notebook with a gradio interface'